refactor: migrate WebAuthn logic to new structure and consolidate components
