+++ src/login/pages/login-username/Page.tsx
@@ -6,7 +6,7 @@ import { clsx } from "keycloakify/tools/clsx";
 import { useState } from "react";
 
 import { Field, FieldError, FieldLabel } from "@/components/ui/field";
-import { WebAuthnConditionalUI } from '@/login/components/WebAuthn/WebAuthConditionalUI';
+import { WebAuthnConditionalUI } from '@/login/components/WebAuthnConditionalUi';
 import { useI18n } from "@/login/i18n";
 import { useKcContext } from "@/login/KcContext";
 import { kcSanitize } from "keycloakify/lib/kcSanitize";
