+++ src/login/pages/webauthn-register/useLogic.ts
@@ -1,4 +1,8 @@
+import { useI18n } from "@/login/i18n";
+import { useKcContext } from "@/login/KcContext";
+import { useRef } from "react";
 import { base64url } from "rfc4648";
+import { assert } from "tsafe/assert";
 
 // see https://github.com/keycloak/keycloak/blob/main/themes/src/main/resources/theme/base/login/resources/js/webauthnRegister.js
 
@@ -14,7 +18,7 @@ export type RegisterOptions = {
     requireResidentKey: string | undefined; // 'Yes' | 'No' | 'not specified'
     userVerificationRequirement?: string;
     createTimeout: number;
-    excludeCredentialIds?: string; // Comma-separated string
+    excludeCredentialIds: string | undefined; // Comma-separated string
     errmsg: string | undefined;
 };
 
@@ -28,7 +32,77 @@ export type WebAuthnRegisterResult =
       }
     | { success: false; error: string };
 
-export async function webAuthnRegister(
+export function useLogic() {
+    const { kcContext } = useKcContext();
+    assert(kcContext.pageId === "webauthn-register.ftl");
+
+    const { msgStr } = useI18n();
+
+    const registerFormRef = useRef<HTMLFormElement>(null);
+
+    const submitRegister = (result: WebAuthnRegisterResult, label?: string) => {
+        const form = registerFormRef.current;
+        assert(form !== null);
+
+        const getInput = (name: string) => {
+            const input = form.elements.namedItem(name);
+            assert(input instanceof HTMLInputElement, `Missing hidden input: ${name}`);
+            return input;
+        };
+
+        if (result.success) {
+            getInput("clientDataJSON").value = result.clientDataJSON;
+            getInput("attestationObject").value = result.attestationObject;
+            getInput("publicKeyCredentialId").value = result.publicKeyCredentialId;
+            getInput("transports").value = result.transports;
+            if (label) getInput("authenticatorLabel").value = label;
+        } else {
+            getInput("error").value = result.error;
+        }
+
+        form.submit();
+    };
+
+    const onRegisterClick = async () => {
+        const result = await webAuthnRegister({
+            challenge: kcContext.challenge,
+            userid: kcContext.userid,
+            username: kcContext.username,
+            signatureAlgorithms: kcContext.signatureAlgorithms,
+            rpEntityName: kcContext.rpEntityName,
+            rpId: kcContext.rpId,
+            attestationConveyancePreference: kcContext.attestationConveyancePreference,
+            authenticatorAttachment: kcContext.authenticatorAttachment,
+            requireResidentKey: kcContext.requireResidentKey,
+            userVerificationRequirement: kcContext.userVerificationRequirement,
+            createTimeout:
+                typeof kcContext.createTimeout === "string"
+                    ? Number(kcContext.createTimeout)
+                    : kcContext.createTimeout,
+            excludeCredentialIds: kcContext.excludeCredentialIds,
+            errmsg: msgStr("webauthn-unsupported-browser-text")
+        });
+
+        if (result.success) {
+            const initLabel = msgStr("webauthn-registration-init-label");
+            const initLabelPrompt = msgStr("webauthn-registration-init-label-prompt");
+
+            let label = window.prompt(initLabelPrompt, initLabel);
+            if (label === null) label = initLabel;
+
+            submitRegister(result, label);
+        } else {
+            submitRegister(result);
+        }
+    };
+
+    return {
+        registerFormRef,
+        onRegisterClick
+    };
+}
+
+async function webAuthnRegister(
     options: RegisterOptions
 ): Promise<WebAuthnRegisterResult> {
     const {
